requirements.txt
absl-py==0.9.0
appdirs==1.4.4
astor==0.8.1
autopep8==1.5.3
Babel==2.8.0
certifi==2020.6.20
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.2.1
cma==3.0.3
colorlog==4.1.0
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
docopt==0.6.2
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Cors==3.0.8
funcsigs==1.0.2
gast==0.3.3
graphviz==0.14
grpcio==1.30.0
gunicorn==20.0.4
hdfs==2.5.8
identify==1.4.20
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
nltk==3.5
nodeenv==1.4.0
numpy==1.19.0
objgraph==3.4.1
opencv-python==4.2.0.34
paddlehub==1.7.1
paddlepaddle==1.8.2
paddleslim==1.0.1
pandas==1.0.5
paddlex==1.1.0
parl==1.3.2
pathlib==1.0.1
Pillow==7.1.2
pre-commit==2.5.1
prettytable==0.7.2
protobuf==3.12.2
psutil==5.7.0
pyarrow==0.13.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==18.0.1
rarfile==3.1
regex==2020.6.8
requests==2.24.0
scipy==1.3.1
sentencepiece==0.1.92
six==1.15.0
tb-nightly==1.15.0a20190801
tensorboardX==1.8
termcolor==1.1.0
toml==0.10.1
tqdm==4.47.0
urllib3==1.25.9
virtualenv==20.0.25
visualdl==2.0.0b7
waitress==1.4.4
Werkzeug==1.0.1
yapf==0.26.0
zipp==3.1.0